backend/pyproject.toml
[project]
name = "subtitle-creator-backend"
version = "1.0.0"
description = "Video subtitle creation backend with Whisper AI"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pysubs2>=1.6.1",
    "ffmpeg-python>=0.2.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "openai-whisper>=20231117",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
